Change keyword substitution from SVN to Git.
Remove the $URL$ keyword, because it is N/A for Git.
Add .gitattributes with "ident" for files with $Id$,
to replace their SVN Id with the file's Git SHA.
This prevents those files from looking like they
came from SVN, and shows their version in Git,
when they are checked out from Git.