wait for defaulted JMX_RMI_PORT to close.
If JMX_RMI_PORT is not specified in app.parameters, then ‘app stop’
did not wait for it to close.  Normally it closes anyway, by the time 
that the other ports do, which were waited on.  But, sometimes it
does not, and causes an immediate ‘app start’ to fail.

Now, ‘app stop’ waits for a defaulted JMX_RMI_PORT to close, too.
If it wasn’t defaulted, then that port is waited on twice, but that
is not a problem, as the second wait takes no time, because the first
has already succeeded.  So, it was not worth complicating the code,
to avoid that duplication.