TAPS-959 widen BAD server startup regexp.
This is to recognize the following problem, which was when Tomcat
was restarted too soon after it was killed, without waiting for
its ports to free up.  Tomcat keeps running, and the context startup
may succeed, but the reverse proxy cannot connect to Tomcat.

SEVERE: Failed to initialize end point associated with ProtocolHandler ["http-bio-10.17.133.22-9677"]
java.net.BindException: Address already in use (Bind failed) /10.17.133.22:9677

The looser pattern is also to allow for different logging
patterns, such as on Frank’s Tenure app.

Note that TAPS normally gets SEVERE errors when it stops, for loose 
threads/memory leaks, which are not really a problem, because we 
always stop Tomcat and the JVM too, not just the TAPS context.  
But sapp-common always rotates the logs after stopping (and before 
starting, if necessary), so any SEVERE errors on stopping will not be 
mistaken for startup problems.