TAPS-959 validate pidfile.
The tapstrngd .server file was missing the "d" on CATALINA_PID and/or PIDFile,
which caused problems, because the scripts and/or systemd did not recognize
that Tomcat had started, but Tomcat continued running.  So, this will be better,
to catch config problems like that, and provide useful error messages, instead
of starting a runaway Tomcat process.