TAPS-959 fix ps_tomcat for $APP_USER.
Match on whole words, to prevent stop() for tapsd on its73 from 
complaining that Tomcat is still running and cannot be killed.  
It was mistaking the Tomcat process of tapsdevd for that of tapsd.  
Tomcat stopped normally, so the pidfile was gone (as normal), 
but then stop() double-checks by using ps_tomcat $APP_USER.  
The tapsd user was matching tapsdevd, failing to kill that process 
(luckily, since that user had no right to do so), and exiting with a 1,
which would make systemd consider the stop to have failed.

This will also be more correct for ps_tomcat $PID, although having
the same substring in the PID of another Tomcat process seems
highly unlikely.

Also, anchor the grep for $APP_USER to the start of the ps line,
to prevent any mismatch in the command line, such as in file names.