MID-609 update for RHEL8.
This is to support the migration of the UH Groupings UI and API apps
to RHEL8, OpenJDK 17, and Tomcat 10.  In particular, the RHEL8 hosts
(iwa74, iws74, iwa04, and iws04) configuration of SELinux (by default?)
does not allow systemd to read the sapp script nor pidfile under the
daemon’s home directory.  This seems to be a change from RHEL7,
and TI would rather not configure SELinux on RHEL8 to be compatible
with RHEL7, so instead, the scripts and app daemon configuration
needs to change for RHEL8.

One fundamental change is that systemd will run sapp-common directly
from /usr/local/bin/appDaemon, and so the daemon user cannot override
a link to it in its home directory and customize that script.
Because of that, the scripts will always load app.parameters
from $HOME, instead of from the same directory containing the script.
And, the scripts will use $(whoami), instead of letting app.parameters
confirm that the correct $APP_USER is trying to run the script.

Whether we fork these scripts at this point, or TI goes back and
updates the configuration of our RHEL7 hosts to be compatible with
these RHEL8 scripts, is TBD.

Other changes:
* move the prep and deploy scripts into the
app script, since the app script is no longer called by SysV init.
* make exit codes more compatible with systemctl.
* add higher-level and positive actions (stop0, stopped,
restart, and provision), to allow simpler Jenkins deploy jobs.