MID-609 use --user instance of systemd on RHEL8.
On RHEL8, this allows better containment of the app to the daemon user's
home directory, like it is on RHEL7 and RHEL6, allows better separation
of concerns, and requires less sudo permissions.  Otherwise, using
systemd --system on RHEL8, SELinux does not allow the sapp script/link
nor Tomcat pidfile to be within the daemon user's home directory.

This commit also allows the same scripts to run on both RHEL7
(with systemctl --system) and RHEL8 (with systemctl --user).
The RHEL7 config for systemctl and sudo can remain unchanged on
our RHEL7 hosts, because RHEL7 does not support systemctl --user.
It is unfortunate that those configs differ between RHEL7 and RHEL8,
but --user seems better and worth using on RHEL8.
This complicates the scripts a little, but it's worth it for RHEL7
to get the new app script commands (e.g., provision and stop0),
consistent with RHEL8.  It is also worth it for RHEL7 to get sapp
flushing its progress messages out of its stderr buffer, so that
the last one on an error will appear in the systemctl status output.