MID-609 add param to exclude catalina.out.
On sapp start, while it is checking catalina.out for a good
context startup, it tails and follows catalina.out, and adds
that output to app-err.log, and tees it to systemctl.  So, it
is displayed by the app script for commands that include start,
and will take up most of the last output lines displayed by
app status and/or systemctl status.  So, it is also included
in the Jenkins job console output and build email.

Frank does not want those lines from catalina.out, in some cases.
The contents of catalina.out when a context starts depend on the app,
and the frameworks upon which the app is built.  This was a change to the
scripts that I made for the RHEL7 migration to systemd, adopting
the context startup check for all apps, which TAPS has been doing
for many years.

So, this commit adds a new param:

EXCLUDE_CATALINA_QUOTE_FROM_APP_ERR_LOG="exclude"

to app.parameters (not set by default) to exclude catalina.out
from app-err.log and systemctl.  Individual apps can do this by
including or uncommenting that param in app.parameters.
In any case, sapp still monitors catalina.out for a good context startup,
and catalina.out still has all of its contents, so we can go look at it there.