TAPS-1004 start checks for .war file and its deployment in Tomcat.
Have the 'app start' check that the expected .war file is in tomcat/webapps,
and that that .war file finishes deploying when Tomcat starts up.
Before this change, if the tomcat/webapps directory is empty, then
the 'app start' will start Tomcat with no webapp and claim to succeed.
Tomcat 7 & 8.5, at least, still say that they deploy the .war, even when
it starts again, after the expanded .war directory is already there.