TAPS-1004 stop using APP_GRP.
The APP_GRP parameter is obsolete, since all of the scripts must be
run by the daemon user.  That was since they started using $HOME and $(whoami)
instead of the APP_HOME and APP_USER parameters respectively.
Now APP_GRP is ignored, and can/should be removed from app.parameters
to avoid confusion.

The APP_GRP was for the original usage of the prep script,
to be run by the developer or SA, not the daemon user, to simplify
the copying of the .war file.  That .war file could remain owned
by the original user, but by prep changing its group to $APP_GRP, the
daemon user could deploy it.  However, now the .war file's group doesn't
matter to these scripts, because the copy made by the prep script
(now app-common) is always owned by the daemon user, because all
the scripts must be run by the daemon user now.

Still changing the .war file's group is normally not a problem,
but certain daemons were not in their APP_GRP, so it became a problem.
For example, on its72 and its73, fdropd and tenured were not in the
fdrop and tenure groups, respectively.  This caused the prep script
to fail, which causes the whole deploy to fail, for no good reason,
until those deploy jobs started ignoring the status code returned
by prep.  This seems like a problem with the system configuration
on its72 and its73 for those daemon users and groups, but the
shared-app-daemon-scripts do not need to enforce configurations
that they do not require.  It seems better to not fail prep for that,
and remove one more unnecessary config param, and allow the
fdrop and tenure deploy jobs to stop ignoring the prep status code.