TAPS-1004 switch to local dbdump script.
Instead of ssh to web12 to run dbdump, run the dbdump.sh from the
deployed webapp locally.  Add a dbdump action for the Jenkins deploy
job to be able to call, especially for after the app has stopped.
Don't include dbdump in deploy, because TAPS needs to do it after
it starts (after it's written the config files for dbdump.sh).

The dbdump config files will not have been created before the app starts,
either for the first time, or since an 'app deploy'.  The dbdump.sh will
fail in that case.