TAPS-1004 update to current app usage, and remove zm.
This removes the obsolete
15-minute Zabbix maintenance from the stop action.  That was
to avoid an old TAPS Down (MIS) alert to a TAPS mailing list.
But, the rule for that alert has been disabled in Zabbix for 
who-knows-how-long, so all the
zm does now is suppress any of the other Zabbix alerts (disk space,
reboots, etc) for its73/33 for that 15-minute period.
That is inconsistent with our other apps, and surprising,
so I’m removing it.  We depend on the Icinga alerts for
all of our apps now, and Zabbix is owned by the SA’s.  If we need
a Zabbix alert for TAPS again, in addition to Icinga, then
we should discuss it with other groups.

This also removes the check and sleep 20 from the stop action,
which was guarding the zm.  The stop action’s success/failure
return status is the same as before, which is the opposite of
‘app stop’ (i.e., failure if the app is stopped, for backwards
compatibility with SysV init).  That usage is inconsistent with
the app script, and confusing, so deprecated.

Also added stop0, provision, and status actions, for consistency with 
the current usage of the app-common script.  The deploy action
is deprecated because it’s basically provision, not ‘app deploy’.
There is a slight change: provision starts with a single follow
of app-err.log across the prep, check or stop, deploy, and start,
instead of separate follows on prep, check, stop, and start,
but it should provide the same output (without the extra 
“Following app-err.log…” messages).

The deploy and stop actions are deprecated, because they are
inconsistent with the same actions in app-common.  They should be
removed after all of the TAPS deploy jobs on all branches
have been using provision and stop0 instead.  The check action is
also deprecated because status is the preferred term with systemctl.